remove full paths from linker options
in order to use the search directory paths instead.

Most windows users end up using the mingw installer when setting up their toolchain, so all these paths end up having to be edited.

This change makes it so that only the search directories need to be changed when the paths are inevitably wrong, and makes setting them in the Win32 build redundant.